Replace shared Interop.FormatMessage.cs with non-shared one
